pyproject.toml
[project]
name = "taskwarrior_scheduler"
version = "0.1.0"
readme = "README.md"
description = "Indicates which taskwarrior's task should be done, considering time spent on other tasks"
authors = [
    {name = "Flávio Tapajós",email = "tapajos@outlook.com"}
]
license = {text = "GPL-3.0-or-later"}
requires-python = ">=3.13"
dependencies = [
    "tasklib (>=2.5.1,<3.0.0)",
    "typer (>=0.15.2,<0.16.0)",
]

[tool.poetry.scripts]
next = "taskwarrior_scheduler.main:app"
next-stats = "taskwarrior_scheduler.stats:app"

[tool.poetry.group.test.dependencies]
pytest-cov = "^6.2.1"
isort = "^6.0.1"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
